Extracting rule for thread count waiting to a new class, with parameterized timeout value. Avoiding the long sleep with a loop.
