+++ driver/src/test/java/org/kaazing/k3po/driver/internal/RobotAgronaThreadsIT.java
@@ -15,31 +15,18 @@
  */
 package org.kaazing.k3po.driver.internal;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.junit.Assert.assertEquals;
 import static org.junit.rules.RuleChain.outerRule;
 
-import java.net.DatagramPacket;
-import java.net.DatagramSocket;
-import java.net.InetSocketAddress;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Set;
-import java.util.regex.Matcher;
-
-import org.hamcrest.Matchers;
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.DisableOnDebug;
 import org.junit.rules.ExpectedException;
 import org.junit.rules.TestRule;
 import org.junit.rules.Timeout;
-import org.junit.runner.Description;
-import org.junit.runners.model.Statement;
 import org.kaazing.k3po.driver.internal.test.utils.K3poTestRule;
 import org.kaazing.k3po.driver.internal.test.utils.TestSpecification;
+import org.kaazing.k3po.driver.internal.test.utils.ThreadCountTestRule;
 
 public class RobotAgronaThreadsIT {
 
@@ -53,40 +40,7 @@ public class RobotAgronaThreadsIT {
         thrown.expect(RuntimeException.class);
     }
     
-    private TestRule countingThreads = new TestRule() {
-        @Override
-        public Statement apply(Statement base, Description description) {
-            return new Statement() {
-                @Override
-                public void evaluate() throws Throwable {
-                    Set<Thread> threadsBefore = Thread.getAllStackTraces().keySet();
-                    int threadCountBefore = threadsBefore.size();
-                    base.evaluate();
-
-                    // let the driver shutdown all threads
-                    Thread.sleep(1000);
-                    Set<Thread> threadsAfter = Thread.getAllStackTraces().keySet();
-                    int threadCountAfter = threadsAfter.size();
-                    
-                    if (threadCountBefore != threadCountAfter) {
-                        System.out.println("Threads before execution:");
-                        for (Iterator<Thread> iterator = threadsBefore.iterator(); iterator.hasNext();) {
-                            Thread type = iterator.next();
-                            System.out.println(type.getName());
-                        }
-                        
-                        System.out.println("\nThreads after execution:");
-                        for (Iterator<Thread> iterator = threadsAfter.iterator(); iterator.hasNext();) {
-                            Thread type = iterator.next();
-                            System.out.println(type.getName());
-                        }
-                    }
-                    
-                    assertEquals("Number of threads is not equal", threadCountBefore, threadCountAfter);
-                }
-            };
-        }
-    };
+    private TestRule countingThreads = new ThreadCountTestRule(1000);
 
 
     @Rule
