+++ driver/src/test/java/org/kaazing/k3po/driver/internal/RobotUdpServerThreadsIT.java
@@ -23,60 +23,23 @@ import static org.junit.rules.RuleChain.outerRule;
 import java.net.DatagramPacket;
 import java.net.DatagramSocket;
 import java.net.InetSocketAddress;
-import java.util.Iterator;
-import java.util.Set;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.DisableOnDebug;
 import org.junit.rules.TestRule;
 import org.junit.rules.Timeout;
-import org.junit.runner.Description;
-import org.junit.runners.model.Statement;
 import org.kaazing.k3po.driver.internal.test.utils.K3poTestRule;
 import org.kaazing.k3po.driver.internal.test.utils.TestSpecification;
+import org.kaazing.k3po.driver.internal.test.utils.ThreadCountTestRule;
 
 public class RobotUdpServerThreadsIT {
 
     private final K3poTestRule robot = new K3poTestRule().setScriptRoot("org/kaazing/k3po/driver/internal/udp");
-    
+
     private final TestRule timeout = new DisableOnDebug(new Timeout(10, SECONDS));
-    
-    private TestRule countingThreads = new TestRule() {
-        @Override
-        public Statement apply(Statement base, Description description) {
-            return new Statement() {
-                @Override
-                public void evaluate() throws Throwable {
-                    Set<Thread> threadsBefore = Thread.getAllStackTraces().keySet();
-                    int threadCountBefore = threadsBefore.size();
-                    base.evaluate();
 
-                    // let the driver shutdown all threads
-                    Thread.sleep(1000);
-                    Set<Thread> threadsAfter = Thread.getAllStackTraces().keySet();
-                    int threadCountAfter = threadsAfter.size();
-                    
-                    if (threadCountBefore != threadCountAfter) {
-                        System.out.println("Threads before execution:");
-                        for (Iterator<Thread> iterator = threadsBefore.iterator(); iterator.hasNext();) {
-                            Thread type = iterator.next();
-                            System.out.println(type.getName());
-                        }
-                        
-                        System.out.println("\nThreads after execution:");
-                        for (Iterator<Thread> iterator = threadsAfter.iterator(); iterator.hasNext();) {
-                            Thread type = iterator.next();
-                            System.out.println(type.getName());
-                        }
-                    }
-                    
-                    assertEquals("Number of threads is not equal", threadCountBefore, threadCountAfter);
-                }
-            };
-        }
-    };
-
+    private TestRule countingThreads = new ThreadCountTestRule(1000);
 
     @Rule
     public final TestRule chain = outerRule(countingThreads).around(robot).around(timeout);
